pyproject.toml
[project]
name = "vibe-plant-website"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dj-database-url>=3.0.1",
    "django>=5.2.5",
    "django-cors-headers>=4.7.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]
